Feat: user can signup consuming api
- Add call to api to persist data
- Add validation check for form data
- Add loader for UI/ UX experience

- Update date of birth validation

[Delivers #165562980]
