+++ README.md
@@ -23,8 +23,8 @@ deposit money..
 
 <b>Api Url:</b>
 
-[Hosted API on Heroku](https://benisi-banka.herokuapp.com/)
-[API documentation](https://benisi-banka.herokuapp.com/swagger/swagger.yaml)
+[Hosted API on Heroku](https://benisi-banka.herokuapp.com/)<br/>
+[API documentation](https://benisi-banka.herokuapp.com/swagger/swagger.json)
 
 ## Features
