+++ UI/account-list.html
@@ -124,6 +124,7 @@
         </footer>
     </div>
     <script src="js/index.js"></script>
+    <script src="js/helper.js"></script>
     <script src="js/account-list.js"></script>
 </body>
 </html>
