+++ UI/account-record.html
@@ -74,6 +74,7 @@
         </footer>
     </div>
     <script src="js/index.js"></script>
+    <script src="js/helper.js"></script>
     <script src="js/account-record.js"></script>
 </body>
 </html>
