+++ UI/bank-account-profile.html
@@ -72,5 +72,6 @@
         </footer>
     </div>
     <script src="js/index.js"></script>
+    <script src="js/helper.js"></script>
 </body>
 </html>
