+++ UI/create-admin-staff-account.html
@@ -82,6 +82,7 @@
         </footer>
     </div>
     <script src="js/index.js"></script>
+    <script src="js/helper.js"></script>
     <script src="create-admin-staff-account.html"></script>
 </body>
 </html>
