+++ UI/transaction-history.html
@@ -169,6 +169,7 @@
         </footer>
     </div>
     <script src="js/index.js"></script>
+    <script src="js/helper.js"></script>
     <script src="js/transactionHistory.js"></script>
 </body>
 </html>
